Move examples to relm4-examples
