--- relm4-examples/examples/popover.rs
@@ -1,83 +1,0 @@
-use gtk::prelude::{BoxExt, ButtonExt, GtkWindowExt, OrientableExt, PopoverExt};
-use relm4::{gtk, send, AppUpdate, Model, RelmApp, Sender, WidgetPlus, Widgets};
-
-#[derive(Default)]
-struct AppModel {
-    counter: u8,
-}
-
-enum AppMsg {
-    Increment,
-    Decrement,
-}
-
-impl Model for AppModel {
-    type Msg = AppMsg;
-    type Widgets = AppWidgets;
-    type Components = ();
-}
-
-impl AppUpdate for AppModel {
-    fn update(&mut self, msg: AppMsg, _components: &(), _sender: Sender<AppMsg>) -> bool {
-        match msg {
-            AppMsg::Increment => {
-                self.counter = self.counter.wrapping_add(1);
-            }
-            AppMsg::Decrement => {
-                self.counter = self.counter.wrapping_sub(1);
-            }
-        }
-        true
-    }
-}
-
-#[relm4::widget]
-impl Widgets<AppModel, ()> for AppWidgets {
-    view! {
-        gtk::ApplicationWindow {
-            set_title: Some("Popup example"),
-            set_default_width: 300,
-            set_default_height: 100,
-            set_child = Some(&gtk::Box) {
-                set_orientation: gtk::Orientation::Vertical,
-                set_margin_all: 5,
-                set_spacing: 5,
-
-                append = &gtk::MenuButton {
-                    set_label: "Pop up!",
-                    set_direction: gtk::ArrowType::Right,
-                    set_popover: popover = Some(&gtk::Popover) {
-                        set_position: gtk::PositionType::Right,
-
-                        set_child = Some(&gtk::Box) {
-                            set_orientation: gtk::Orientation::Vertical,
-                            set_spacing: 5,
-                            append = &gtk::Button {
-                                set_label: "Increment counter",
-                                connect_clicked(sender) => move |_| {
-                                    send!(sender, AppMsg::Increment);
-                                }
-                            },
-                            append = &gtk::Button {
-                                set_label: "Decrement counter",
-                                connect_clicked(sender) => move |_| {
-                                    send!(sender, AppMsg::Decrement);
-                                }
-                            },
-                        },
-                    },
-                },
-                append = &gtk::Label {
-                    set_margin_all: 5,
-                    set_label: watch! { &format!("Counter: {}", model.counter) },
-                }
-            },
-        }
-    }
-}
-
-fn main() {
-    let model = AppModel::default();
-    let app = RelmApp::new(model);
-    app.run();
-}
